move useFetch hook folder
